fix(config): switch to float32 — MPS float16 NaN issue with pixel art model

+++ src/pixelforge/config.py
@@ -36,7 +36,7 @@ class GenerationConfig(BaseModel):
     palette_colors: int = 24
 
     device: str = "mps"
-    dtype: str = "float16"
+    dtype: str = "float32"
 
 
 DEFAULT_CONFIG = GenerationConfig()
